+++ modules/game.py
@@ -1,0 +1,62 @@
+from modules.info import GameState, WHITE, BLACK
+from modules.ui_element import UIElement
+import pygame
+
+def play_level(screen):
+    return_btn = UIElement(
+        center_position=(160, 570),
+        font_size=20,
+        bg_rgb=None,
+        text_rgb=WHITE,
+        text="Return to main menu",
+        action=GameState.TITLE,
+    )
+
+    while True:
+        mouse_up = False
+        for event in pygame.event.get():
+            if event.type == pygame.MOUSEBUTTONUP and event.button == 1:
+                mouse_up = True
+        screen.fill(BLACK)
+
+        ui_action = return_btn.update(pygame.mouse.get_pos(), mouse_up)
+        if ui_action is not None:
+            return ui_action
+        return_btn.draw(screen)
+
+        pygame.display.flip()
+
+def title_screen(screen):
+    start_btn = UIElement(
+        center_position=(500, 400),
+        font_size=30,
+        bg_rgb=None,
+        text_rgb=WHITE,
+        text="Start",
+        action=GameState.NEWGAME,
+    )
+    quit_btn = UIElement(
+        center_position=(500, 500),
+        font_size=30,
+        bg_rgb=None,
+        text_rgb=WHITE,
+        text="Quit",
+        action=GameState.QUIT,
+    )
+
+    buttons = [start_btn, quit_btn]
+
+    while True:
+        mouse_up = False
+        for event in pygame.event.get():
+            if event.type == pygame.MOUSEBUTTONUP and event.button == 1:
+                mouse_up = True
+        screen.fill(BLACK)
+
+        for button in buttons:
+            ui_action = button.update(pygame.mouse.get_pos(), mouse_up)
+            if ui_action is not None:
+                return ui_action
+            button.draw(screen)
+
+        pygame.display.flip()
