+++ modules/info.py
@@ -1,5 +1,11 @@
 from enum import Enum
 
+RED = (255, 0, 0)
+GREEN = (0, 255, 0)
+BLUE = (0, 0, 255)
+WHITE = (255, 255, 255)
+BLACK = (0, 0, 0)
+
 class GameState(Enum):
     QUIT = -1
     TITLE = 0
